refactor: move insight merchant ranking types
