+++ frontend/src/insights/types/merchantRanking.ts
@@ -1,0 +1,8 @@
+export type MerchantRankingRow = {
+  id: string
+  name: string
+  totalAmount: number
+  transactionCount: number
+  averageAmount: number
+  changePct: number | null
+}
